Fix/Suppress some mypy and prospector warnings

+++ telegram_periodic_msg_bot/message/message_deleter.py
@@ -51,11 +51,12 @@ class MessageDeleter:
     def DeleteMessage(self,
                       message: pyrogram.types.Message) -> bool:
         try:
-            self.client.delete_messages(message.chat.id, PyrogramWrapper.MessageId(message))
-            return True
+            if message.chat is not None:
+                self.client.delete_messages(message.chat.id, PyrogramWrapper.MessageId(message))
+                return True
         except pyrogram_ex.forbidden_403.MessageDeleteForbidden:
             self.logger.GetLogger().exception(f"Unable to delete message {PyrogramWrapper.MessageId(message)}")
-            return False
+        return False
 
     # Delete messages
     def DeleteMessages(self,
